@@ -5,9 +5,48 @@ import AJAX from "../ajax";
 import Flash from "../flash";
 
 import "../styles/pages/box-show.scss";
+import Modal from "../modal";
 
-$(document).ready(() => {
+$(() => {
     getBoxTrackingMovements();
+    getBoxInCrate();
+
+    $(`.edit-box`).click(() => {
+        const ajax = AJAX.route(`POST`, `box_edit_template`, {
+            box: $('#box-id').val()
+        });
+        Modal.load(ajax, {
+            success : () => {
+                window.location.href = Routing.generate(`box_show`, {
+                    box: $('#box-id').val()
+                });
+            }
+        })
+    });
+
+    $(`.delete-box`).click(() => {
+        const ajax = AJAX.route(`POST`, `box_delete_template`, {
+            box: $('#box-id').val()
+        });
+        Modal.load(ajax, {
+            success : () =>{
+                window.location.href = Routing.generate(`boxes_list`);
+            }
+        })
+    });
+
+    $(document).arrive('.delete-box-in-crate', function() {
+        $(this).click(function() {
+            const ajax = AJAX.route(`POST`, `box_remove_crate_template`, {
+                box: $(this).data('id'),
+                crate: $('#box-id').val()
+            });
+            Modal.load(ajax, {
+                success: (response) => {
+                    $('.refresh-after-add').replaceWith(response.template);
+                }});
+            });
+    })
 
     $('.comment-search').on('change', function () {
         getBoxTrackingMovements(0);
@@ -29,6 +68,20 @@ $(document).ready(() => {
         }, 800);
         return false;
     });
+
+    $('.add-box-to-cart').on(`change`, function (){
+        const box = $(this).val();
+
+        const ajax = AJAX.route(`GET`, `box_add_crate`, {
+            box: box,
+            crate: $('#box-id').val()
+        });
+        ajax.json()
+            .then(response =>{
+                $('.refresh-after-add').replaceWith(response.template);
+            });
+        $(this).empty();
+    });
 });
 
 function getBoxTrackingMovements(start = 0) {
@@ -39,27 +92,47 @@ function getBoxTrackingMovements(start = 0) {
 
     const search = $('.comment-search').val();
 
-    AJAX.route('GET', 'get_box_mouvements', {box: $('#box-id').val(), search, start})
-        .json((result) => {
+    AJAX.route('GET', 'box_movements', {box: $('#box-id').val(), search, start})
+        .json()
+        .then((result) => {
             if(result.success) {
                 if (start === 0) {
                     $('.history-wrapper').empty();
                 }
                 const data = (result.data || []);
-                const historyLines = data.map(({state, color, comment, date}) => {
-                    const $rawComment = $($.parseHTML(comment));
-                    const trimmedComment = $rawComment.text().trim();
+                const historyLines = data.map(({isCurrentRecord, state, crate, quality, date, time, operator, location, depository}) => {
+                    const $rawQuality = $($.parseHTML(quality));
+                    const trimmedQuality = $rawQuality.text().trim();
 
-                    let $comment = `<div class="timeline-line-comment"></div>`;
-                    if(trimmedComment.trim()) {
-                        $comment = `<div class="timeline-line-comment alert alert-${color}">${comment}</div>`;
+                    let $quality = `<div class="timeline-line-comment"></div>`;
+                    if(trimmedQuality.trim()) {
+                        $quality = `<div class="timeline-line-comment">${quality}</div>`;
+                    }
+
+                    let subtitle;
+                    if(operator){
+                        subtitle = `Opérateur : ${operator}`;
+                    }
+                    if (depository) {
+                        subtitle = subtitle ? `${subtitle} - ` : '';
+                        subtitle += depository;
+                    }
+                    if (location) {
+                        subtitle = subtitle ? `${subtitle} - ` : '';
+                        subtitle += location;
+                    }
+
+                    if (crate) {
+                        state += ` dans la caisse <a href="${Routing.generate('box_show', {box: crate.id})}">${crate.number}</a>`;
                     }
 
                     return `
-                        <div class="timeline-line">
-                            <span class="timeline-line-marker"><strong>${date}</strong></span>
-                            <span class="timeline-line-title ml-3">${state}</span>
-                            ${$comment}
+                        <div class="timeline-line d-flex ${isCurrentRecord ? 'current-timeline-line' : ''}">
+                            <span class="timeline-line-marker"><strong>${date}</strong><p>${time}</p></span>
+                            <div class="timeline-line-title ml-3">
+                                <div class="d-flex"><strong>${state}</strong>${$quality}</div>
+                                <p>${subtitle}</p>
+                            </div>
                         </div>
                     `;
                 });
@@ -96,4 +169,14 @@ function getBoxTrackingMovements(start = 0) {
                 $showMoreWrapper.remove();
             }
         })
+}
+
+function getBoxInCrate() {
+    const ajax = AJAX.route(`GET`, `box_in_crate_api`, {
+        id: $('#box-id').val()
+    });
+    ajax.json()
+        .then(response =>{
+            $('.refresh-after-add').replaceWith(response.template);
+        });
 }
