@@ -3,16 +3,17 @@
 namespace App\Controller\Referential;
 
 use App\Annotation\HasPermission;
+use App\Controller\AbstractController;
 use App\Entity\BoxType;
 use App\Entity\GlobalSetting;
 use App\Entity\Role;
 use App\Helper\Form;
 use App\Helper\FormatHelper;
 use App\Repository\BoxTypeRepository;
+use App\Service\BoxTypeService;
 use App\Service\ExportService;
 use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -30,6 +31,7 @@ class BoxTypeController extends AbstractController {
         $settingsRepository = $manager->getRepository(GlobalSetting::class);
         $capacities = explode(",", $settingsRepository->getValue(GlobalSetting::BOX_CAPACITIES));
         $shapes = explode(",", $settingsRepository->getValue(GlobalSetting::BOX_SHAPES));
+        $boxTypeRepository = $manager->getRepository(BoxType::class);
 
         return $this->render("referential/box_type/index.html.twig", [
             "new_box_type" => new BoxType(),
@@ -37,6 +39,7 @@ class BoxTypeController extends AbstractController {
             "box_types_order" => BoxTypeRepository::DEFAULT_DATATABLE_ORDER,
             "capacities" => $capacities ?: [],
             "shapes" => $shapes ?: [],
+            "starterKit" => $boxTypeRepository->findStarterKit(),
         ]);
     }
 
@@ -74,13 +77,17 @@ class BoxTypeController extends AbstractController {
      * @Route("/nouveau", name="box_type_new", options={"expose": true})
      * @HasPermission(Role::MANAGE_BOX_TYPES)
      */
-    public function new(Request $request, EntityManagerInterface $manager): Response {
+    public function new(Request $request,
+                        EntityManagerInterface $entityManager,
+                        BoxTypeService $boxTypeService): Response {
         $form = Form::create();
 
         $content = (object)$request->request->all();
-        $existing = $manager->getRepository(BoxType::class)->findOneBy(["name" => $content->name]);
+        $content->image = $request->files->get('image');
+
+        $existing = $entityManager->getRepository(BoxType::class)->findOneBy(["name" => $content->name]);
         if ($existing) {
-            $form->addError("name", "Ce type de Box existe déjà");
+            $form->addError("name", "Ce type de Box / Caisse existe déjà");
         }
 
         if ($content->price < 0) {
@@ -89,18 +96,12 @@ class BoxTypeController extends AbstractController {
 
         if ($form->isValid()) {
             $boxType = new BoxType();
-            $boxType->setName($content->name)
-                ->setPrice($content->price)
-                ->setActive($content->active)
-                ->setCapacity($content->capacity)
-                ->setShape($content->shape);
-
-            $manager->persist($boxType);
-            $manager->flush();
+            $boxTypeService->persistBoxType($entityManager, $boxType, $content);
+            $entityManager->flush();
 
             return $this->json([
                 "success" => true,
-                "message" => "Type de Box créé avec succès",
+                "message" => "Type de Box / Caisse créé avec succès",
             ]);
         } else {
             return $form->errors();
@@ -115,12 +116,13 @@ class BoxTypeController extends AbstractController {
         $settingsRepository = $manager->getRepository(GlobalSetting::class);
         $capacities = explode(",", $settingsRepository->getValue(GlobalSetting::BOX_CAPACITIES));
         $shapes = explode(",", $settingsRepository->getValue(GlobalSetting::BOX_SHAPES));
-
+        $boxTypeRepository = $manager->getRepository(BoxType::class);
         return $this->json([
             "submit" => $this->generateUrl("box_type_edit", ["boxType" => $boxType->getId()]),
             "template" => $this->renderView("referential/box_type/modal/edit.html.twig", [
                 "box_type" => $boxType,
                 "capacities" => $capacities ?: [],
+                "starterKit" => $boxTypeRepository->findStarterKit(),
                 "shapes" => $shapes ?: [],
             ])
         ]);
@@ -130,27 +132,27 @@ class BoxTypeController extends AbstractController {
      * @Route("/modifier/{boxType}", name="box_type_edit", options={"expose": true})
      * @HasPermission(Role::MANAGE_BOX_TYPES)
      */
-    public function edit(Request $request, EntityManagerInterface $manager, BoxType $boxType): Response {
+    public function edit(Request $request,
+                         EntityManagerInterface $entityManager,
+                         BoxTypeService $boxTypeService,
+                         BoxType $boxType): Response {
         $form = Form::create();
 
         $content = (object)$request->request->all();
-        $existing = $manager->getRepository(BoxType::class)->findOneBy(["name" => $content->name]);
+        $content->image = $request->files->get('image');
+
+        $existing = $entityManager->getRepository(BoxType::class)->findOneBy(["name" => $content->name]);
         if ($existing !== null && $existing !== $boxType) {
-            $form->addError("name", "Un autre type de Box avec ce nom existe déjà");
+            $form->addError("name", "Un autre type de Box / Caisse avec ce nom existe déjà");
         }
 
         if ($form->isValid()) {
-            $boxType->setName($content->name)
-                ->setPrice($content->price)
-                ->setActive($content->active)
-                ->setCapacity($content->capacity)
-                ->setShape($content->shape);
-
-            $manager->flush();
+            $boxTypeService->persistBoxType($entityManager, $boxType, $content);
+            $entityManager->flush();
 
             return $this->json([
                 "success" => true,
-                "message" => "Type de Box modifié avec succès",
+                "message" => "Type de Box / Caisse modifié avec succès",
             ]);
         } else {
             return $form->errors();
