@@ -3,6 +3,7 @@
 namespace App\Controller\Referential;
 
 use App\Annotation\HasPermission;
+use App\Controller\AbstractController;
 use App\Entity\Group;
 use App\Entity\Role;
 use App\Helper\Form;
@@ -10,7 +11,6 @@ use App\Repository\GroupRepository;
 use App\Service\ExportService;
 use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
