@@ -11,9 +11,9 @@ use App\Security\Authenticator;
 use App\Service\Mailer;
 use DateTime;
 use LogicException;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\Encoder\NativePasswordEncoder;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
@@ -52,7 +52,7 @@ class SecurityController extends AbstractController {
     /**
      * @Route("/register", name="register")
      */
-    public function register(Request $request, Mailer $mailer, UserPasswordEncoderInterface $encoder): Response {
+    public function register(Request $request, Mailer $mailer, UserPasswordHasherInterface $encoder): Response {
         if ($this->getUser()) {
             return $this->redirectToRoute(Authenticator::HOME_ROUTE);
         }
@@ -70,6 +70,9 @@ class SecurityController extends AbstractController {
 
             $user->setEmail($request->request->get("email"))
                 ->setUsername($request->request->get("username"))
+                ->setDeliverer(false)
+                ->setDeliveryAssignmentMail(false)
+                ->setDeliveryAssignmentPreparationMail(false)
                 ->addGroup($em->getRepository(Group::class)->find($request->request->get("group")));
 
             if ($existing) {
@@ -93,7 +96,7 @@ class SecurityController extends AbstractController {
             if ($valid) {
                 $noAccess = $em->getRepository(Role::class)->findOneBy(["code" => Role::ROLE_NO_ACCESS]);
 
-                $user->setPassword($encoder->encodePassword($user, $password))
+                $user->setPassword($encoder->hashPassword($user, $password))
                     ->setActive(true)
                     ->setRole($noAccess)
                     ->setCreationDate(new DateTime());
