@@ -3,19 +3,19 @@
 namespace App\Controller\Settings;
 
 use App\Annotation\HasPermission;
+use App\Controller\AbstractController;
 use App\Entity\Role;
 use App\Entity\User;
 use App\Helper\Form;
 use App\Helper\FormatHelper;
-use WiiCommon\Helper\StringHelper;
 use App\Repository\RoleRepository;
 use App\Service\ExportService;
 use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use WiiCommon\Helper\StringHelper;
 
 /**
  * @Route("/parametrage/roles")
@@ -92,7 +92,8 @@ class RoleController extends AbstractController {
                 ->setActive($content->active)
                 ->setPermissions(explode(",", $content->permissions))
                 ->setAllowEditOwnGroupOnly($content->allowEditOwnGroupOnly)
-                ->setShowNewOrderOnHome($content->showNewOrderOnHome)
+                ->setRedirectNewCounterOrder($content->redirectNewCounterOrder)
+                ->setShowIndicatorsOnHome($content->showIndicatorsOnHome)
                 ->setReceiveMailsNewAccounts($content->receiveMailsNewAccounts);
 
             if(count($role->getPermissions()) === 0) {
@@ -165,7 +166,8 @@ class RoleController extends AbstractController {
                 ->setActive($content->active)
                 ->setPermissions(explode(",", $content->permissions))
                 ->setAllowEditOwnGroupOnly($content->allowEditOwnGroupOnly)
-                ->setShowNewOrderOnHome($content->showNewOrderOnHome)
+                ->setRedirectNewCounterOrder($content->redirectNewCounterOrder)
+                ->setShowIndicatorsOnHome($content->showIndicatorsOnHome)
                 ->setReceiveMailsNewAccounts($content->receiveMailsNewAccounts);
 
             if(count($role->getPermissions()) === 0) {
