@@ -13,29 +13,7 @@ use Doctrine\ORM\Mapping as ORM;
  */
 class Box {
 
-    public const AVAILABLE = 1;
-    public const UNAVAILABLE = 2;
-    public const CONSUMER = 3;
-    public const CLIENT = 4;
-    public const OUT = 5;
-
-    public const NAMES = [
-        self::AVAILABLE => "Disponible",
-        self::UNAVAILABLE => "Indisponible",
-        self::CONSUMER => "Consommateur",
-        self::CLIENT => "Client",
-        self::OUT => "Sorti",
-    ];
-
-    public const LINKED_COLORS = [
-        self::AVAILABLE => "success",
-        self::UNAVAILABLE => "danger",
-        self::CONSUMER => "primary",
-        self::CLIENT => "warning",
-        self::OUT => "secondary",
-    ];
-
-    public const DEFAULT_COLOR = 'dark';
+    public const OWNER_BOXEATY = "boxeaty";
 
     /**
      * @ORM\Id
@@ -95,24 +73,72 @@ class Box {
     private ?int $uses = null;
 
     /**
+     * @ORM\Column(type="boolean")
+     */
+    private ?bool $isBox = null;
+
+    /**
      * @ORM\OneToMany(targetEntity=DepositTicket::class, mappedBy="box")
      */
     private Collection $depositTickets;
 
     /**
-     * @ORM\ManyToMany(targetEntity=Order::class, mappedBy="boxes")
+     * @ORM\ManyToMany(targetEntity=CounterOrder::class, mappedBy="boxes")
      */
-    private Collection $orders;
+    private Collection $counterOrders;
 
     /**
      * @ORM\Column(type="datetime", nullable=false)
      */
-    private DateTime $creationDate;
+    private ?DateTime $creationDate;
+
+    /**
+     * @ORM\ManyToMany(targetEntity=Collect::class, mappedBy="crates")
+     */
+    private Collection $collects;
+
+    /**
+     * @ORM\OneToMany(targetEntity=Box::class, mappedBy="crate")
+     */
+    private Collection $containedBoxes;
+
+    /**
+     * @ORM\ManyToOne(targetEntity=Box::class, inversedBy="containedBoxes")
+     */
+    private ?Box $crate = null;
+
+    /**
+     * @ORM\OneToMany(targetEntity=PreparationLine::class, mappedBy="crate")
+     */
+    private Collection $cratePreparationLines;
+
+    /**
+     * @ORM\ManyToMany(targetEntity=PreparationLine::class, mappedBy="boxes")
+     */
+    private Collection $boxPreparationLines;
+
+    /**
+     * @ORM\OneToMany(targetEntity=BoxRecord::class, mappedBy="crate")
+     */
+    private Collection $cratePackingRecords;
 
     public function __construct() {
         $this->boxRecords = new ArrayCollection();
-        $this->orders = new ArrayCollection();
-        $this->creationDate = new DateTime('now');
+        $this->counterOrders = new ArrayCollection();
+        $this->creationDate = new DateTime("now");
+        $this->collects = new ArrayCollection();
+        $this->containedBoxes = new ArrayCollection();
+        $this->cratePreparationLines = new ArrayCollection();
+        $this->boxPreparationLines = new ArrayCollection();
+        $this->cratePackingRecords = new ArrayCollection();
+    }
+
+    public function fromRecord(BoxRecord $record): self {
+        return $this->setState($record->getState())
+            ->setLocation($record->getLocation())
+            ->setQuality($record->getQuality())
+            ->setOwner($record->getClient())
+            ->setComment($record->getComment());
     }
 
     public function getId(): ?int {
@@ -286,45 +312,253 @@ class Box {
     }
 
     /**
-     * @return Collection|Order[]
+     * @return Collection|CounterOrder[]
      */
-    public function getOrders(): Collection {
-        return $this->orders;
+    public function getCounterOrders(): Collection {
+        return $this->counterOrders;
     }
 
-    public function addOrder(Order $order): self {
-        if (!$this->orders->contains($order)) {
-            $this->orders[] = $order;
+    public function addCounterOrder(CounterOrder $order): self {
+        if (!$this->counterOrders->contains($order)) {
+            $this->counterOrders[] = $order;
             $order->addBox($this);
         }
 
         return $this;
     }
 
-    public function removeOrder(Order $order): self {
-        if ($this->orders->removeElement($order)) {
+    public function removeCounterOrder(CounterOrder $order): self {
+        if ($this->counterOrders->removeElement($order)) {
             $order->removeBox($this);
         }
 
         return $this;
     }
 
-    public function fromRecord(BoxRecord $record): self {
-        return $this->setState($record->getState())
-            ->setLocation($record->getLocation())
-            ->setQuality($record->getQuality())
-            ->setOwner($record->getClient())
-            ->setComment($record->getComment());
-    }
 
-    public function getCreationDate(): ?\DateTimeInterface
-    {
+    public function getCreationDate(): ?DateTime {
         return $this->creationDate;
     }
 
-    public function setCreationDate(\DateTimeInterface $creationDate): self
-    {
+    public function setCreationDate(DateTime $creationDate): self {
         $this->creationDate = $creationDate;
+
+        return $this;
+    }
+
+    /**
+     * @return Collection|Collect[]
+     */
+    public function getCollects(): Collection {
+        return $this->collects;
+    }
+
+    public function addCollect(Collect $collect): self {
+        if (!$this->collects->contains($collect)) {
+            $this->collects[] = $collect;
+            $collect->addCrate($this);
+        }
+
+        return $this;
+    }
+
+    public function removeCollect(Collect $collect): self {
+        if ($this->collects->removeElement($collect)) {
+            $collect->removeCrate($this);
+        }
+
+        return $this;
+    }
+
+    public function setCollect(?array $collects): self {
+        foreach ($this->getCollects()->toArray() as $collect) {
+            $this->removeCollect($collect);
+        }
+
+        $this->collects = new ArrayCollection();
+        foreach ($collects as $collect) {
+            $this->addCollect($collect);
+        }
+
+        return $this;
+    }
+
+    public function isBox(): ?bool {
+        return $this->isBox;
+    }
+
+    public function setIsBox(?bool $isBox): self {
+        $this->isBox = $isBox;
+
+        return $this;
+    }
+
+    public function getCrate(): ?Box{
+        return $this->crate;
+    }
+
+    public function setCrate(?Box $crate): self {
+        if($this->crate && $this->crate !== $crate) {
+            $this->crate->removeContainedBox($this);
+        }
+        $this->crate = $crate;
+        if($crate) {
+            $crate->addContainedBox($this);
+        }
+
+        return $this;
+    }
+
+    /**
+     * @return Collection|Box[]
+     */
+    public function getContainedBoxes(): Collection {
+        return $this->containedBoxes;
+    }
+
+    public function addContainedBox(Box $containedBox): self {
+        if (!$this->containedBoxes->contains($containedBox)) {
+            $this->containedBoxes[] = $containedBox;
+            $containedBox->setCrate($this);
+        }
+
+        return $this;
+    }
+
+    public function removeContainedBox(Box $containedBox): self {
+        if ($this->containedBoxes->removeElement($containedBox)) {
+            if ($containedBox->getCrate() === $this) {
+                $containedBox->setCrate(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function setContainedBoxes(?array $containedBoxes): self {
+        foreach($this->getContainedBoxes()->toArray() as $containedBox) {
+            $this->removeContainedBox($containedBox);
+        }
+
+        $this->containedBoxes = new ArrayCollection();
+        foreach($containedBoxes as $containedBox) {
+            $this->addContainedBox($containedBox);
+        }
+
+        return $this;
+    }
+
+    /**
+     * @return Collection|PreparationLine[]
+     */
+    public function getCratePreparationLines(): Collection {
+        return $this->cratePreparationLines;
+    }
+
+    public function addCratePreparationLine(PreparationLine $line): self {
+        if (!$this->cratePreparationLines->contains($line)) {
+            $this->cratePreparationLines[] = $line;
+            $line->setCrate($this);
+        }
+
+        return $this;
+    }
+
+    public function removeCratePreparationLine(PreparationLine $line): self {
+        if ($this->cratePreparationLines->removeElement($line)) {
+            if ($line->getCrate() === $this) {
+                $line->setCrate(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function setCratePreparationLines(?array $lines): self {
+        foreach($this->getCratePreparationLines()->toArray() as $line) {
+            $this->removeCratePreparationLine($line);
+        }
+
+        $this->cratePreparationLines = new ArrayCollection();
+        foreach($lines as $line) {
+            $this->addCratePreparationLine($line);
+        }
+
+        return $this;
+    }
+
+    /**
+     * @return Collection|Box[]
+     */
+    public function getBoxPreparationLines(): Collection {
+        return $this->boxPreparationLines;
+    }
+
+    public function addBoxPreparationLine(PreparationLine $line): self {
+        if (!$this->boxPreparationLines->contains($line)) {
+            $this->boxPreparationLines[] = $line;
+            $line->addBox($this);
+        }
+
+        return $this;
+    }
+
+    public function removeBoxPreparationLine(PreparationLine $line): self {
+        if ($this->boxPreparationLines->removeElement($line)) {
+            $line->removeBox($this);
+        }
+
+        return $this;
+    }
+
+    public function setBoxPreparationLines(?array $lines): self {
+        foreach($this->getBoxPreparationLines()->toArray() as $line) {
+            $this->removeBoxPreparationLine($line);
+        }
+
+        $this->boxPreparationLines = new ArrayCollection();
+        foreach($lines as $line) {
+            $this->addBoxPreparationLine($line);
+        }
+
+        return $this;
+    }
+
+    /**
+     * @return Collection|BoxRecord[]
+     */
+    public function getCratePackingRecords(): Collection {
+        return $this->cratePackingRecords;
+    }
+
+    public function addCratePackingRecord(BoxRecord $boxRecord): self {
+        if (!$this->cratePackingRecords->contains($boxRecord)) {
+            $this->cratePackingRecords[] = $boxRecord;
+            $boxRecord->setCrate($this);
+        }
+
+        return $this;
+    }
+
+    public function removeCratePackingRecord(BoxRecord $boxRecord): self {
+        if ($this->cratePackingRecords->removeElement($boxRecord)) {
+            if ($boxRecord->getCrate() === $this) {
+                $boxRecord->setCrate(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function setCratePackingRecords(?array $boxRecords): self {
+        foreach($this->getCratePackingRecords()->toArray() as $boxRecord) {
+            $this->removeCratePackingRecord($boxRecord);
+        }
+
+        $this->cratePackingRecords = new ArrayCollection();
+        foreach($boxRecords as $boxRecord) {
+            $this->addCratePackingRecord($boxRecord);
+        }
 
         return $this;
     }
