@@ -2,6 +2,7 @@
 
 namespace App\Entity;
 
+use App\Entity\Utils\ActiveTrait;
 use App\Repository\GroupRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
@@ -13,7 +14,7 @@ use Doctrine\ORM\Mapping as ORM;
  */
 class Group {
 
-    use Active;
+    use ActiveTrait;
 
     /**
      * @ORM\Id
