@@ -2,20 +2,35 @@
 
 namespace App\Entity;
 
-use WiiCommon\Helper\Stream;
+use App\Entity\Utils\ActiveTrait;
 use App\Repository\LocationRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
+use WiiCommon\Helper\Stream;
 
 /**
  * @ORM\Entity(repositoryClass=LocationRepository::class)
  */
 class Location {
 
-    use Active;
+    use ActiveTrait;
 
     public const MIN_KIOSK_CAPACITY = 1;
+
+    const RECEPTION = "Reception";
+    const EXPEDITION = "Expedition";
+    const STOCK = "Stock";
+    const QUALITY = "Qualite";
+    const CLIENT = "Client";
+
+    const LOCATION_TYPES = [
+        1 => self::RECEPTION,
+        2 => self::EXPEDITION,
+        3 => self::STOCK,
+        4 => self::QUALITY,
+        5 => self::CLIENT,
+    ];
 
     /**
      * @ORM\Id
@@ -27,7 +42,7 @@ class Location {
     /**
      * @ORM\ManyToOne(targetEntity=Location::class)
      */
-    private ?Location $deporte = null;
+    private ?Location $offset = null;
 
     /**
      * @ORM\Column(type="boolean")
@@ -45,9 +60,14 @@ class Location {
     private ?string $description = null;
 
     /**
-     * @ORM\ManyToOne(targetEntity=Client::class, inversedBy="kiosks")
+     * @ORM\ManyToOne(targetEntity=Client::class, inversedBy="locations")
      */
     private ?Client $client = null;
+
+    /**
+     * @ORM\OneToOne(targetEntity=Client::class, mappedBy="outLocation")
+     */
+    private ?Client $outClient = null;
 
     /**
      * @ORM\Column(type="integer", nullable=true)
@@ -60,7 +80,7 @@ class Location {
     private ?string $message = null;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="integer", nullable=true)
      */
     private ?int $deposits = null;
 
@@ -79,21 +99,43 @@ class Location {
      */
     private Collection $depositTickets;
 
+    /**
+     * @ORM\OneToMany(targetEntity=Collect::class, mappedBy="pickLocation")
+     */
+    private Collection $pickedCollects;
+
+    /**
+     * @ORM\OneToMany(targetEntity=Collect::class, mappedBy="dropLocation")
+     */
+    private Collection $droppedCollects;
+
+    /**
+     * @ORM\Column(type="integer", nullable=true)
+     */
+    private ?int $type = null;
+
+    /**
+     * @ORM\ManyToOne(targetEntity=Depository::class, inversedBy="locations")
+     */
+    private ?Depository $depository = null;
+
     public function __construct() {
         $this->boxes = new ArrayCollection();
         $this->boxRecords = new ArrayCollection();
+        $this->pickedCollects = new ArrayCollection();
+        $this->droppedCollects = new ArrayCollection();
     }
 
     public function getId(): ?int {
         return $this->id;
     }
 
-    public function getDeporte(): ?Location {
-        return $this->deporte;
+    public function getOffset(): ?Location {
+        return $this->offset;
     }
 
-    public function setDeporte(?Location $deporte): self {
-        $this->deporte = $deporte;
+    public function setOffset(?Location $offset): self {
+        $this->offset = $offset;
         return $this;
     }
 
@@ -133,6 +175,15 @@ class Location {
     public function setClient(?Client $client): self {
         $this->client = $client;
 
+        return $this;
+    }
+
+    public function getOutClient(): ?Client {
+        return $this->outClient;
+    }
+
+    public function setOutClient(?Client $outClient): self {
+        $this->outClient = $outClient;
         return $this;
     }
 
@@ -260,5 +311,116 @@ class Location {
                 ->toArray()
         ];
     }
+
+    /**
+     * @return Collection|Collect[]
+     */
+    public function getPickedCollects(): Collection {
+        return $this->pickedCollects;
+    }
+
+    public function addPickedCollect(Collect $pickedCollect): self {
+        if (!$this->pickedCollects->contains($pickedCollect)) {
+            $this->pickedCollects[] = $pickedCollect;
+            $pickedCollect->setPickLocation($this);
+        }
+
+        return $this;
+    }
+
+    public function removePickedCollect(Collect $pickedCollects): self {
+        if ($this->pickedCollects->removeElement($pickedCollects)) {
+            // set the owning side to null (unless already changed)
+            if ($pickedCollects->getPickLocation() === $this) {
+                $pickedCollects->setPickLocation(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function setPickedCollect(?array $pickedCollects): self {
+        foreach ($this->getPickedCollects()->toArray() as $pickedCollect) {
+            $this->removePickedCollect($pickedCollect);
+        }
+
+        $this->pickedCollects = new ArrayCollection();
+        foreach ($pickedCollects as $pickedCollect) {
+            $this->addPickedCollect($pickedCollect);
+        }
+
+        return $this;
+    }
+
+    /**
+     * @return Collection|Collect[]
+     */
+    public function getDroppedCollects(): Collection {
+        return $this->droppedCollects;
+    }
+
+    public function addDroppedCollect(Collect $droppedCollect): self {
+        if (!$this->droppedCollects->contains($droppedCollect)) {
+            $this->droppedCollects[] = $droppedCollect;
+            $droppedCollect->setDropLocation($this);
+        }
+
+        return $this;
+    }
+
+    public function removeDroppedCollect(Collect $droppedCollect): self {
+        if ($this->droppedCollects->removeElement($droppedCollect)) {
+            // set the owning side to null (unless already changed)
+            if ($droppedCollect->getDropLocation() === $this) {
+                $droppedCollect->setDropLocation(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function setDroppedCollects(?array $droppedCollects): self {
+        foreach ($this->getDroppedCollects()->toArray() as $droppedCollect) {
+            $this->removeDroppedCollect($droppedCollect);
+        }
+
+        $this->droppedCollects = new ArrayCollection();
+        foreach ($droppedCollects as $droppedCollect) {
+            $this->addDroppedCollect($droppedCollect);
+        }
+
+        return $this;
+    }
+
+
+    public function getType(): ?int
+    {
+        return $this->type;
+    }
+
+    public function setType(?int $type): self
+    {
+        $this->type = $type;
+
+        return $this;
+    }
+
+    public function getDepository(): ?Depository
+    {
+        return $this->depository;
+    }
+
+    public function setDepository(?Depository $depository): self {
+        if($this->depository && $this->depository !== $depository) {
+            $this->depository->removeLocation($this);
+        }
+        $this->depository = $depository;
+        if($depository) {
+            $depository->addLocation($this);
+        }
+
+        return $this;
+    }
+
 
 }
