@@ -4,11 +4,15 @@ namespace App\Listener;
 
 use App\Annotation\Authenticated;
 use App\Annotation\HasPermission;
+use App\Controller\AbstractController;
+use App\Entity\User;
 use App\Service\RoleService;
 use Doctrine\Common\Annotations\AnnotationReader;
+use Doctrine\ORM\EntityManagerInterface;
 use ReflectionClass;
 use ReflectionException;
 use RuntimeException;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController as SymfonyAbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\ControllerArgumentsEvent;
@@ -23,8 +27,11 @@ class AnnotationListener {
     /** @Required */
     public Environment $templating;
 
+    /** @Required */
+    public EntityManagerInterface $manager;
+
     public function onRequest(ControllerArgumentsEvent $event) {
-        if (!$event->isMasterRequest() || !is_array($event->getController())) {
+        if (!$event->isMainRequest() || !is_array($event->getController())) {
             return;
         }
 
@@ -47,6 +54,11 @@ class AnnotationListener {
         if ($annotation instanceof HasPermission) {
             $this->handleHasPermission($event, $annotation);
         }
+
+        $annotation = $reader->getMethodAnnotation($method, Authenticated::class);
+        if ($annotation instanceof Authenticated) {
+            $this->handleAuthenticated($event, $controller);
+        }
     }
 
     private function handleAuthenticated(ControllerArgumentsEvent $event) {
@@ -62,7 +74,7 @@ class AnnotationListener {
 
     private function handleHasPermission(ControllerArgumentsEvent $event, HasPermission $annotation) {
         if (!$this->roleService->hasPermission(...$annotation->value)) {
-            $event->setController(function() use ($annotation) {
+            $event->setController(function () use ($annotation) {
                 if ($annotation->mode == HasPermission::IN_JSON) {
                     return new JsonResponse([
                         "success" => false,
@@ -74,6 +86,26 @@ class AnnotationListener {
                     throw new RuntimeException("Unknown mode $annotation->mode");
                 }
             });
+        }
+    }
+
+    private function handleAuthenticated(ControllerArgumentsEvent $event, SymfonyAbstractController $controller) {
+        $request = $event->getRequest();
+
+        if (!($controller instanceof AbstractController)) {
+            throw new RuntimeException("Routes annotated with @Authenticated must extend App\\Controller\\AbstractController");
+        }
+
+        $userRepository = $this->manager->getRepository(User::class);
+
+        $authorization = $request->headers->get("x-authorization", "");
+        preg_match("/Bearer (\w*)/i", $authorization, $matches);
+
+        $user = $matches ? $userRepository->findByApiKey($matches[1]) : null;
+        if ($user) {
+            $controller->setUser($user);
+        } else {
+            throw new UnauthorizedHttpException("no challenge");
         }
     }
 
