@@ -6,6 +6,8 @@ use App\Entity\Box;
 use App\Entity\BoxRecord;
 use App\Entity\User;
 use App\Helper\QueryHelper;
+use App\Service\BoxStateService;
+use Doctrine\Common\Collections\Criteria;
 use Doctrine\ORM\EntityRepository;
 
 /**
@@ -29,7 +31,7 @@ class BoxRecordRepository extends EntityRepository {
             ->addSelect("record.state AS state")
             ->addSelect("client.name AS client_name")
             ->addSelect("user.username AS user_username")
-            ->where("record.trackingMovement = 1")
+            ->andWhere("record.trackingMovement = 1")
             ->leftJoin("record.location", "location")
             ->leftJoin("record.box", "box")
             ->leftJoin("record.quality", "quality")
@@ -43,7 +45,7 @@ class BoxRecordRepository extends EntityRepository {
         $search = $params["search"]["value"] ?? null;
 
         $qb = $this->createQueryBuilder("record")
-            ->where('record.trackingMovement = 1');
+            ->andWhere('record.trackingMovement = 1');
 
         QueryHelper::withCurrentGroup($qb, "record.client.group", $user);
 
@@ -90,7 +92,7 @@ class BoxRecordRepository extends EntityRepository {
             }
         }
 
-        if (!empty($params['order'])) {
+        if (!empty($params["order"])) {
             foreach ($params["order"] ?? [] as $order) {
                 $column = $params["columns"][$order["column"]]["data"];
                 $qb->addOrderBy("record.$column", $order["dir"]);
@@ -116,7 +118,7 @@ class BoxRecordRepository extends EntityRepository {
 
     public function findPreviousTrackingMovement(Box $box): ?BoxRecord {
         return $this->createQueryBuilder("record")
-            ->where("record.box = :box")
+            ->andWhere("record.box = :box")
             ->andWhere("record.location IS NOT NULL")
             ->andWhere("record.trackingMovement = 1")
             ->orderBy("record.id", "DESC")
@@ -127,25 +129,50 @@ class BoxRecordRepository extends EntityRepository {
     }
 
     public function getBoxRecords(Box $box, int $start, int $length, string $search = null): array {
-        $qb = $this->createQueryBuilder("record")
-            ->select("record.comment AS comment")
+        $queryBuilder = $this->createQueryBuilder("record");
+
+        $exprBuilder = $queryBuilder->expr();
+
+        $queryBuilder
+            ->select("record.id AS id")
+            ->addSelect("join_quality.name AS quality")
             ->addSelect("record.date AS date")
             ->addSelect("record.state AS state")
-            ->where("record.box = :box")
-            ->andWhere("record.trackingMovement = 0")
+            ->addSelect("join_operator.username AS operator")
+            ->addSelect("join_location.name AS location")
+            ->addSelect("join_depository.name AS depository")
+            ->addSelect("join_crate.number AS crateNumber")
+            ->addSelect("join_crate.id AS crateId")
+            ->leftJoin("record.user", "join_operator")
+            ->leftJoin("record.quality", "join_quality")
+            ->leftJoin("record.location", "join_location")
+            ->leftJoin("join_location.depository", "join_depository")
+            ->leftJoin("record.crate", "join_crate")
+            ->andWhere("record.box = :box")
+            ->andWhere($exprBuilder->orX(
+                "record.trackingMovement = 0",
+                "record.state IN (:packingStates)"
+            ))
             ->orderBy("record.date", "DESC")
             ->addOrderBy("record.id", "DESC")
-            ->setParameter("box", $box);
+            ->setParameter("box", $box)
+            ->setParameter("packingStates", [BoxStateService::STATE_RECORD_PACKING, BoxStateService::STATE_RECORD_UNPACKING]);
 
         if($search) {
-            $qb
-                ->andWhere("record.comment LIKE :search")
+            $queryBuilder
+                ->andWhere($exprBuilder->orX(
+                    "join_quality.name LIKE :search",
+                    "join_location.name LIKE :search",
+                    "join_depository.name LIKE :search",
+                    "join_operator.username LIKE :search",
+                    "DATE(record.date) LIKE :search"
+                ))
                 ->setParameter("search", '%' . $search . '%');
         }
 
         return [
-            'totalCount' => QueryHelper::count($qb, 'record'),
-            'data' => $qb
+            "totalCount" => QueryHelper::count($queryBuilder, "record"),
+            "data" => $queryBuilder
                 ->setMaxResults($length)
                 ->setFirstResult($start)
                 ->getQuery()
@@ -153,21 +180,19 @@ class BoxRecordRepository extends EntityRepository {
         ];
     }
 
-    public function findNewerTrackingMovement(BoxRecord $trackingMovement): ?BoxRecord {
-        $box = $trackingMovement->getBox();
-        if ($box
-            && $trackingMovement->getId()
-            && $trackingMovement->getDate()) {
+    public function findNewerTrackingMovement(BoxRecord $record): ?BoxRecord {
+        $box = $record->getBox();
+        if ($box && $record->getId() && $record->getDate()) {
             return $this->createQueryBuilder("record")
-                ->where("record.box = :box")
+                ->andWhere("record.box = :box")
                 ->andWhere("record.id != :movement")
                 ->andWhere("record.date > :date")
                 ->andWhere("record.trackingMovement = 1")
                 ->addOrderBy("record.date", "DESC")
                 ->addOrderBy("record.id", "DESC")
                 ->setParameter("box", $box)
-                ->setParameter("movement", $trackingMovement->getId())
-                ->setParameter("date", $trackingMovement->getDate())
+                ->setParameter("movement", $record->getId())
+                ->setParameter("date", $record->getDate())
                 ->setMaxResults(1)
                 ->getQuery()
                 ->getOneOrNullResult();
@@ -175,4 +200,35 @@ class BoxRecordRepository extends EntityRepository {
 
         return null;
     }
+
+    public function getNumberBoxByStateAndDate( \DateTime $startDate, \DateTime $endDate, int $state, array $clients) {
+        return $this->createQueryBuilder("record")
+            ->select("COUNT(DISTINCT record.id) AS result")
+            ->andWhere("record.date BETWEEN :dateMin AND :dateMax")
+            ->andWhere("record.state = :state")
+            ->andWhere("record.client in (:clients)")
+            ->setParameter("dateMin", $startDate)
+            ->setParameter("dateMax", $endDate)
+            ->setParameter("state", $state)
+            ->setParameter("clients", $clients)
+            ->getQuery()
+            ->getSingleScalarResult();
+    }
+
+    public function findCurrentBoxRecord(Box $box): ?BoxRecord {
+        return $this->createQueryBuilder("record")
+            ->andWhere("record.box = :box")
+            ->andWhere("record.trackingMovement = 0 OR record.state IN (:packingStates)")
+            ->orderBy("record.date", Criteria::DESC)
+            ->addOrderBy("record.id", Criteria::DESC)
+            ->setMaxResults(1)
+            ->setParameter("box", $box)
+            ->setParameter("packingStates", [
+                BoxStateService::STATE_RECORD_PACKING,
+                BoxStateService::STATE_RECORD_UNPACKING
+            ])
+            ->getQuery()
+            ->getOneOrNullResult();
+    }
+
 }
