@@ -2,9 +2,16 @@
 
 namespace App\Repository;
 
+use App\Entity\Box;
 use App\Entity\BoxType;
+use App\Entity\ClientOrder;
+use App\Entity\Depository;
+use App\Entity\Status;
+use App\Helper\FormatHelper;
 use App\Helper\QueryHelper;
+use App\Service\BoxStateService;
 use Doctrine\ORM\EntityRepository;
+use WiiCommon\Helper\Stream;
 
 /**
  * @method BoxType|null find($id, $lockMode = null, $lockVersion = null)
@@ -36,7 +43,7 @@ class BoxTypeRepository extends EntityRepository {
         $total = QueryHelper::count($qb, "box_type");
 
         if ($search) {
-            $qb->where($qb->expr()->orX(
+            $qb->andWhere($qb->expr()->orX(
                 "box_type.name LIKE :search",
                 "box_type.price LIKE :search",
                 "box_type.capacity LIKE :search",
@@ -45,13 +52,12 @@ class BoxTypeRepository extends EntityRepository {
         }
 
 
-        if (!empty($params['order'])) {
+        if (!empty($params["order"])) {
             foreach ($params["order"] ?? [] as $order) {
                 $column = $params["columns"][$order["column"]]["data"];
                 $qb->addOrderBy("box_type.$column", $order["dir"]);
             }
-        }
-        else {
+        } else {
             foreach (self::DEFAULT_DATATABLE_ORDER as [$column, $dir]) {
                 $qb->addOrderBy("box_type.$column", $dir);
             }
@@ -69,15 +75,125 @@ class BoxTypeRepository extends EntityRepository {
         ];
     }
 
-    public function getForSelect(?string $search) {
-        return $this->createQueryBuilder("box_type")
-            ->select("box_type.id AS id, box_type.name AS text")
-            ->where("box_type.name LIKE :search")
+    public function getForSelect(?string $search, $extended = false) {
+        $boxTypes = $this->createQueryBuilder("box_type")
+            ->andWhere("box_type.name LIKE :search")
             ->andWhere("box_type.active = 1")
             ->setMaxResults(15)
             ->setParameter("search", "%$search%")
             ->getQuery()
-            ->getArrayResult();
+            ->getResult();
+
+        return Stream::from($boxTypes)
+            ->map(fn(BoxType $boxType) => [
+                'id' => $boxType->getId(),
+                'text' => $extended
+                    ? $boxType->getName() . ' - ' . ($boxType->getVolume()
+                        ? $boxType->getVolume() . 'm³'
+                        : 'N/C') . ' - ' . FormatHelper::price($boxType->getPrice())
+                    : $boxType->getName(),
+                'name' => $boxType->getName(),
+                'price' => $boxType->getPrice(),
+                'volume' => $boxType->getVolume(),
+                'image' => $boxType->getImage()
+                    ? $boxType->getImage()->getPath()
+                    : null,
+            ])
+            ->values();
+    }
+
+    public function findStarterKit() {
+        $boxType = $this->createQueryBuilder("box_type")
+            ->andWhere("box_type.name LIKE :kit")
+            ->setParameter("kit", BoxType::STARTER_KIT)
+            ->getQuery()
+            ->getSingleResult();
+
+        $volumeLabel = (
+        $boxType->getVolume()
+            ? $boxType->getVolume() . 'm³'
+            : 'N/C'
+        );
+
+        return [
+            'id' => $boxType->getId(),
+            'text' => $boxType->getName() . ' - ' . $volumeLabel . ' - ' . FormatHelper::price($boxType->getPrice()),
+            'name' => $boxType->getName(),
+            'price' => $boxType->getPrice(),
+            'volume' => $boxType->getVolume(),
+            'image' => $boxType->getImage()
+                ? $boxType->getImage()->getPath()
+                : null,
+        ];
+    }
+
+    public function countAvailableInDepository(Depository $depository, int $defaultCrateType, array $types = []): array {
+        $totalAvailableResult = $this->createQueryBuilder("box_type")
+            ->select("box_type.id AS id")
+            ->addSelect("COUNT(box.id) AS count")
+            ->addSelect("owner.id AS client")
+            ->innerJoin("box_type.boxes", "box")
+            ->leftJoin("box.owner", "owner")
+            ->innerJoin("box.location", "location")
+            ->innerJoin("box.quality", "quality")
+            ->innerJoin("location.depository", "depository")
+            ->andWhere("box.state = :availableState")
+            ->andWhere("depository = :depository")
+            ->andWhere("quality.clean = 1")
+            ->andWhere("box_type IN (:types)")
+            ->setParameter("types", $types)
+            ->setParameter("availableState", BoxStateService::STATE_BOX_AVAILABLE)
+            ->setParameter("depository", $depository)
+            ->groupBy("box_type.id, owner.id")
+            ->getQuery()
+            ->getResult();
+
+        $totalAvailable = [];
+        foreach ($totalAvailableResult as $line) {
+            $totalAvailable[$line["id"]][$line["client"] ?: Box::OWNER_BOXEATY] = $line["count"];
+        }
+        $inUnpreparedResult = $this->createQueryBuilder("box_type")
+            ->select("box_type.id AS id")
+            ->addSelect("SUM(order_lines.quantity) AS count")
+            ->addSelect("order_client.id AS client")
+            ->join("box_type.clientOrderLines", "order_lines")
+            ->join("order_lines.clientOrder", "order")
+            ->join("order.client", "order_client")
+            ->join("order.status", "status")
+            ->andWhere("status.code = :status")
+            ->andWhere("box_type IN (:types)")
+            ->setParameter("types", $types)
+            ->setParameter("status", Status::CODE_ORDER_PLANNED)
+            ->groupBy("box_type.id, order_client.id")
+            ->getQuery()
+            ->getResult();
+
+        if (in_array($defaultCrateType, $types)) {
+            array_push($inUnpreparedResult, ...$this->getEntityManager()
+                ->createQueryBuilder()
+                ->from(ClientOrder::class, "client_order")
+                ->select("$defaultCrateType AS id")
+                ->addSelect("SUM(client_order.cratesAmount) AS count")
+                ->addSelect("client_order_client.id AS client")
+                ->join("client_order.client", "client_order_client")
+                ->join("client_order.status", "status")
+                ->andWhere("status.code = :status")
+                ->setParameter("status", Status::CODE_ORDER_PLANNED)
+                ->groupBy("client_order_client.id")
+                ->getQuery()
+                ->getResult());
+        }
+        $inUnprepared = [];
+        foreach ($inUnpreparedResult as $line) {
+            $owner = $line["client"] ?: Box::OWNER_BOXEATY;
+            $inUnprepared[$line["id"]][$owner] = $line["count"];
+        }
+        foreach ($totalAvailable as $type => $clients) {
+            foreach ($clients as $client => $count) {
+                $totalAvailable[$type][$client] = $count - ($inUnprepared[$type][$client] ?? 0);
+            }
+        }
+        return $totalAvailable;
     }
 
 }
