@@ -5,7 +5,9 @@ namespace App\Repository;
 use App\Entity\Client;
 use App\Entity\User;
 use App\Helper\QueryHelper;
+use DateTime;
 use Doctrine\ORM\EntityRepository;
+use WiiCommon\Helper\Stream;
 
 /**
  * @method Client|null find($id, $lockMode = null, $lockVersion = null)
@@ -46,26 +48,46 @@ class ClientRepository extends EntityRepository {
         if ($search) {
             $qb->leftJoin("client.users", "search_user")
                 ->leftJoin("client.group", "search_group")
+                ->leftJoin("client.contact", "search_contact")
                 ->leftJoin("client.linkedMultiSite", "search_multi_site")
                 ->andWhere($qb->expr()->orX(
                     "client.name LIKE :search",
                     "client.address LIKE :search",
                     "search_user.username LIKE :search",
+                    "search_contact.username LIKE :search",
                     "search_group.name LIKE :search",
                     "search_multi_site.name LIKE :search",
                 ))
                 ->setParameter("search", "%$search%");
         }
 
-        if (!empty($params['order'])) {
+        foreach ($params["filters"] ?? [] as $name => $value) {
+            switch ($name) {
+                case("client"):
+                    $qb
+                        ->andWhere("client.id = :filter_client")
+                        ->setParameter("filter_client", $value);
+                    break;
+                case("multiSite"):
+                    $qb->andWhere("client.linkedMultiSite = :filter_multiSite")
+                        ->setParameter("filter_multiSite", $value);
+                    break;
+                default:
+                    $qb->andWhere("client.$name = :filter_$name")
+                        ->setParameter("filter_$name", $value);
+                    break;
+            }
+        }
+
+        if (!empty($params["order"])) {
             foreach ($params["order"] ?? [] as $order) {
                 $column = $params["columns"][$order["column"]]["data"];
                 if ($column === "contact") {
                     QueryHelper::order($qb, "client.contact.username", $order["dir"]);
                 } else if ($column === "group") {
                     QueryHelper::order($qb, "client.group.name", $order["dir"]);
-                } else if ($column === "multiSite") {
-                    QueryHelper::order($qb, "client.multiSite.name", $order["dir"]);
+                } else if ($column === "linkedMultiSite") {
+                    QueryHelper::order($qb, "client.linkedMultiSite.name", $order["dir"]);
                 } else {
                     $qb->addOrderBy("client.$column", $order["dir"]);
                 }
@@ -89,11 +111,29 @@ class ClientRepository extends EntityRepository {
         ];
     }
 
-    public function getForSelect(?string $search, ?string $group, ?User $user) {
+    public function getForSelect(?string $search,
+                                 ?string $group,
+                                 ?User $user,
+                                 ?bool $costInformationNeeded = false) {
         $qb = $this->createQueryBuilder("client");
 
-        if($user && $user->getRole()->isAllowEditOwnGroupOnly()) {
-            $qb->andWhere("client.group IN (:groups)")
+        $exprBuilder = $qb->expr();
+
+        if ($user && $user->getRole()->isAllowEditOwnGroupOnly()) {
+            $userClients = $user->getClients();
+
+            if (!$userClients->isEmpty()) {
+                $qb
+                    ->andWhere($exprBuilder->orX(
+                        'client IN (:userClients)',
+                        'linkedMultiSite IN (:userClients)'
+                    ))
+                    ->leftJoin('client.linkedMultiSite', 'linkedMultiSite')
+                    ->setParameter('userClients', $userClients);
+            }
+
+            $qb
+                ->andWhere("client.group IN (:groups)")
                 ->setParameter("groups", $user->getGroups());
         }
 
@@ -102,11 +142,29 @@ class ClientRepository extends EntityRepository {
                 ->setParameter("group", $group);
         }
 
-        return $qb->select("client.id AS id, client.name AS text")
+        $qb
+            ->select("client.id AS id")
+            ->addSelect("client.name AS text")
+            ->addSelect("join_information.workingDayDeliveryRate AS workingRate")
+            ->addSelect("join_information.nonWorkingDayDeliveryRate AS nonWorkingRate")
+            ->addSelect("join_information.serviceCost AS serviceCost")
+            ->addSelect("client.address AS address")
+            ->addSelect("join_deliveryMethod.id AS deliveryMethod")
+            ->leftjoin("client.clientOrderInformation", "join_information")
+            ->leftjoin("join_information.deliveryMethod", "join_deliveryMethod")
             ->andWhere("client.name LIKE :search")
             ->andWhere("client.active = 1")
             ->setMaxResults(15)
-            ->setParameter("search", "%$search%")
+            ->setParameter("search", "%$search%");
+
+        if ($costInformationNeeded) {
+            $qb
+                ->andWhere("join_information.workingDayDeliveryRate IS NOT NULL")
+                ->andWhere("join_information.nonWorkingDayDeliveryRate IS NOT NULL")
+                ->andWhere("join_information.serviceCost IS NOT NULL");
+        }
+
+        return $qb
             ->getQuery()
             ->getArrayResult();
     }
@@ -120,13 +178,49 @@ class ClientRepository extends EntityRepository {
         }
 
         return $qb->select("client.id AS id, client.name AS text")
-            ->where("client.name LIKE :search")
+            ->andWhere("client.name LIKE :search")
             ->andWhere("client.active = 1")
             ->andWhere("client.isMultiSite = 1")
             ->setMaxResults(15)
             ->setParameter("search", "%$search%")
             ->getQuery()
             ->getArrayResult();
+    }
+
+    public function findActiveRecurrence(): array {
+        $now = new DateTime("today midnight");
+
+        return $this->createQueryBuilder("client")
+            ->join("client.clientOrderInformation", "client_order_information")
+            ->join("client.cratePatternLines", "crate_pattern_lines")
+            ->join("client_order_information.deliveryMethod", "delivery_method")
+            ->join("client_order_information.depository", "depository")
+            ->join("client_order_information.orderRecurrence", "recurrence")
+            ->andWhere("client.active = 1")
+            ->andWhere("recurrence.id IS NOT NULL")
+            ->andWhere("recurrence.start <= :now")
+            ->andWhere("recurrence.end >= :now")
+            ->andWhere("crate_pattern_lines.id IS NOT NULL")
+            ->andWhere("delivery_method.id IS NOT NULL")
+            ->andWhere("depository.id IS NOT NULL")
+            ->andWhere("client_order_information.workingDayDeliveryRate IS NOT NULL")
+            ->andWhere("client_order_information.nonWorkingDayDeliveryRate IS NOT NULL")
+            ->andWhere("DAY(recurrence.lastEdit) = DAY(:now)")
+            ->andWhere("MONTH(recurrence.lastEdit) = MONTH(:now)")
+            ->setParameter("now", $now)
+            ->getQuery()
+            ->getResult();
+    }
+
+    public function getCratePatternAmountGroupedByClient(): array {
+        $clients = $this->createQueryBuilder("client")
+            ->andWhere('client.cratePatternLines IS NOT EMPTY')
+            ->getQuery()
+            ->getResult();
+
+        return Stream::from($clients)
+            ->keymap(fn (Client $client) => [$client->getId(), $client->getCratePatternAmount()])
+            ->toArray();
     }
 
 }
