@@ -7,6 +7,7 @@ use App\Entity\User;
 use App\Helper\QueryHelper;
 use DateTime;
 use Doctrine\ORM\EntityRepository;
+use WiiCommon\Helper\Stream;
 
 /**
  * @method DepositTicket|null find($id, $lockMode = null, $lockVersion = null)
@@ -98,7 +99,7 @@ class DepositTicketRepository extends EntityRepository {
             }
         }
 
-        if (!empty($params['order'])) {
+        if (!empty($params["order"])) {
             foreach ($params["order"] ?? [] as $order) {
                 $column = $params["columns"][$order["column"]]["data"];
                 if ($column === "kiosk") {
@@ -167,6 +168,23 @@ class DepositTicketRepository extends EntityRepository {
             ->setParameter("valid", DepositTicket::VALID)
             ->getQuery()
             ->getResult();
+    }
+
+    public function countByStatusGroupedByType(string $status): array {
+        $res = $this->createQueryBuilder("deposit_ticket")
+            ->select("COUNT(deposit_ticket.id) AS count")
+            ->addSelect('join_box_type.id AS type')
+            ->join("deposit_ticket.box", "join_box")
+            ->join("join_box.type", "join_box_type")
+            ->andWhere("deposit_ticket.state = :status")
+            ->groupBy('join_box_type.id')
+            ->setParameter("status", $status)
+            ->getQuery()
+            ->getResult();
+
+        return Stream::from($res)
+            ->keymap(fn (array $line) => [$line['type'], $line['count']])
+            ->toArray();
     }
 
 }
