@@ -34,11 +34,11 @@ class GroupRepository extends EntityRepository {
         $total = QueryHelper::count($qb, "g");
 
         if ($search) {
-            $qb->where("g.name LIKE :search")
+            $qb->andWhere("g.name LIKE :search")
                 ->setParameter("search", "%$search%");
         }
 
-        if (!empty($params['order'])) {
+        if (!empty($params["order"])) {
             foreach ($params["order"] ?? [] as $order) {
                 $column = $params["columns"][$order["column"]]["data"];
                 $qb->addOrderBy("g.$column", $order["dir"]);
@@ -73,7 +73,7 @@ class GroupRepository extends EntityRepository {
         } else {
             return $this->createQueryBuilder("g")
                 ->select("g.id AS id, g.name AS text")
-                ->where("g.name LIKE :search")
+                ->andWhere("g.name LIKE :search")
                 ->andWhere("g.active = 1")
                 ->setMaxResults(15)
                 ->setParameter("search", "%$search%")
