@@ -3,7 +3,6 @@
 namespace App\Twig;
 
 use App\Helper\FormatHelper;
-use WiiCommon\Helper\Stream;
 use App\Service\RoleService;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
@@ -12,6 +11,7 @@ use Twig\Environment;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFilter;
 use Twig\TwigFunction;
+use WiiCommon\Helper\Stream;
 
 class AppExtension extends AbstractExtension {
 
@@ -78,7 +78,15 @@ class AppExtension extends AbstractExtension {
     }
 
     private function shouldAddItem(array $item): bool {
-        return !isset($item["permission"]) || $this->hasPermission(constant($item["permission"]));
+        if(isset($item["permission"])) {
+            if (is_array($item["permission"])) {
+                return $this->hasPermission(...array_map("constant", $item["permission"]));
+            } else {
+                return $this->hasPermission(constant($item["permission"]));
+            }
+        } else {
+            return true;
+        }
     }
 
     public function hasPermission(string ...$permissions): bool {
